Move TrackRemovalArgs to RemoveTrack.h
